refactoring of autolabor managers, part 1
First step in refactoring labor management plugins.

* `AutoLaborManager` base class for labor managers created

`labormanager` management logic split out into new derived `LaborManagementClassic` implementation

remaining `labormanager' plugin shell modified to act as a supervisor of multiple managers

also, `joblabormapper` changed to throw exceptions on inference failure instead of printing debug messages, mainly for data dependency reasons; the mesage print and pause is now handled by the caller with a try-catch.
