+++ plugins/labormanager/AutoLaborManager.cpp
@@ -1,0 +1,15 @@
+
+#include "AutoLaborManager.h"
+
+#include <Console.h>
+#include <PluginManager.h>
+
+#include <vector>
+
+using namespace DFHack;
+
+command_result AutoLaborManager::do_command(color_ostream& out,
+    std::vector<std::string>& parameters)
+{
+    return CR_WRONG_USAGE;
+}
